#25 Improve around app & home state managers
* Store welcome screen passed flag when `skip` is tapped
* Remove unnecessary `App` repository
